+++ presentation/src/main/java/com/mashup/presentation/feature/login/LoginViewModel.kt
@@ -6,7 +6,7 @@ import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.mashup.domain.exception.ConflictException
-import com.mashup.domain.usecase.*
+import com.mashup.domain.usecase.login.*
 import com.mashup.presentation.ui.common.ValidationState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
